moar

+++ src/test/fuzz/mempool.cpp
@@ -22,21 +22,9 @@ void test_one_input(const std::vector<uint8_t>& buffer)
 {
     FuzzedDataProvider fuzzed_data_provider{buffer.data(), buffer.size()};
     CTxMemPool mpool;
-
-    // AcceptToMemoryPool
-    // removeForBlock
-    // 
-    // need block disconnected / connected callbacks...
-    
+   
     while (fuzzed_data_provider.ConsumeBool()) {
-	    // args to AcceptToMemoryPool or removeForBlock...
-	    // bool AcceptToMemoryPool(CTxMemPool&, TxValidationState&, CTransactionRef&,
-	    // std::list<CTransactionRef>*,bool,CAmount,bool)
-	    // 
-	    // void addUnchecked(const CTxMemPoolEntry&, bool) EXCL(cs, cs_main)
-	    //
-	    // void removeForBlock(const std::vector<CTransactionRef>&,unsigned int) EXCL(cs)
-	switch (fuzzed_data_provider.ConsumeIntegralInRange<int>(0, 1)) {
+	switch (fuzzed_data_provider.ConsumeIntegralInRange<uint8_t>(0, 1)) {
 	case 0: {
 	    std::optional<CMutableTransaction> mtx = ConsumeDeserializable<CMutableTransaction>(fuzzed_data_provider);
 	    if (!mtx) {
@@ -48,19 +36,13 @@ void test_one_input(const std::vector<uint8_t>& buffer)
 	    break;
 	}
 	case 1: {
-	    // LOCK2(cs_main, mpool.cs);
 	    std::optional<CMutableTransaction> mtx = ConsumeDeserializable<CMutableTransaction>(fuzzed_data_provider);
 	    if (!mtx) {
 		break;
 	    }
 	    const CTransaction ctx{*mtx};
-
-	    // Choose a dummy reason
-	    // TODO: can also use removeForBlock with ConsumeBytes<CTransactionRef>?
-	
-	    //auto reason = fuzzed_data_provider.ConsumeEnum<MemPoolRemovalReason>()
 	    MemPoolRemovalReason reason;
-	    switch (fuzzed_data_provider.ConsumeIntegralInRange<int>(0, 5)) {
+	    switch (fuzzed_data_provider.ConsumeIntegralInRange<uint8_t>(0, 5)) {
 	    case 0: {
 	        reason = MemPoolRemovalReason::EXPIRY;
 	    	break;
